Update to new sdk

+++ src/DiskCover/Infrastructure/AimpMessageHook.cs
@@ -4,13 +4,13 @@ using AIMP.SDK.MessageDispatcher;
 
 namespace AIMP.DiskCover.Infrastructure
 {
-    public delegate AimpActionResult OnCoreMessage(AimpCoreMessageType message, int param1, int param2);
+    public delegate AimpActionResult OnCoreMessage(AimpCoreMessageType message, int param1, IntPtr param2);
 
     public class AimpMessageHook : IAimpMessageHook
     {
         public event OnCoreMessage OnCoreMessage;
 
-        public AimpActionResult CoreMessage(AimpCoreMessageType message, int param1, int param2)
+        public AimpActionResult CoreMessage(AimpCoreMessageType message, int param1, IntPtr param2)
         {
             if (OnCoreMessage != null)
             {
